# Conflicts:
#   app/appSettings.js
#   app/controllers/Empresas/EmpresasCompletarController.js
#   app/controllers/ReportesController.js

+++ app/controllers/Promocions/PromocionsUpdateController.js
@@ -1,5 +1,5 @@
 (function () {
-  var PromocionsUpdateController = function ($scope, $log, $location, $cookieStore, $routeParams, PromocionsFactory, FileUploader, AccesosAmazonFactory) {
+  var PromocionsUpdateController = function ($scope, $log, $location, $cookies, $routeParams, PromocionsFactory, FileUploader, AccesosAmazonFactory) {
 
     var IdPromocion = $routeParams.IdPromocion;
     $scope.Promocion = {};
@@ -188,25 +188,27 @@
 
     function eliminarImagen(data) {
       var Url = $scope.Promocion.Url;
-      var resultado = Url.split("/");
-      var picturePath = 'Anexos' + '/' + resultado[5];
-      var s3Client = new AWS.S3({
-        accessKeyId: data[0].AccessKey,
-        secretAccessKey: data[0].SecretAccess,
-        params: {
-          Bucket: data[0].Bucket,
-        },
-      });
-
-      s3Client.deleteObject({
-        Key: picturePath,
-      }, function (err, data) {
-
-      });
+      if (Url) {
+        var resultado = Url.split("/");
+        var picturePath = 'Anexos' + '/' + resultado[5];
+        var s3Client = new AWS.S3({
+          accessKeyId: data[0].AccessKey,
+          secretAccessKey: data[0].SecretAccess,
+          params: {
+            Bucket: data[0].Bucket,
+          },
+        });
+  
+        s3Client.deleteObject({
+          Key: picturePath,
+        }, function (err, data) {
+  
+        });
+      }
     };
   };
 
-  PromocionsUpdateController.$inject = ['$scope', '$log', '$location', '$cookieStore', '$routeParams', 'PromocionsFactory', 'FileUploader', 'AccesosAmazonFactory'];
+  PromocionsUpdateController.$inject = ['$scope', '$log', '$location', '$cookies', '$routeParams', 'PromocionsFactory', 'FileUploader', 'AccesosAmazonFactory'];
 
   angular.module('marketplace').controller('PromocionsUpdateController', PromocionsUpdateController);
 }());
